data_module/src/Data_Processing/requirements.txt
alembic==1.0.2
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
enum34==1.1.6
fdfs-client==4.0.7
Flask==0.10.1
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
Flask-Migrate==2.1.0
Flask-Script==2.0.5
Flask-Session==0.3.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.5
ImageHash==4.0
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.1.0
meld3==1.0.2
MySQL-python==1.2.5
numpy==1.15.4
passlib==1.7.1
pika==0.12.0
Pillow==5.3.0
pycparser==2.19
PyJWT==1.6.4
PyMySQL==0.9.2
pyOpenSSL==18.0.0
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
PyWavelets==1.0.1
redis==2.10.5
requests==2.18.3
scipy==1.1.0
six==1.11.0
SQLAlchemy==1.1.12
supervisor==3.3.4
urllib3==1.22
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
